Refactor HTML structure and scripts across multiple pages: replace inline theme script with external theme-boot.js, streamline navigation elements, and remove redundant footer content. Update playfolio.js script to module type for improved functionality.
