+++ css/playfolio/motion.css
@@ -1,0 +1,125 @@
+/* Motion and portrait surfaces */
+
+.pf-ambassador-lockup {
+  position: absolute;
+  left: 1.25rem;
+  bottom: 1.25rem;
+  z-index: 4;
+  display: inline-grid;
+  grid-template-columns: auto auto;
+  align-items: center;
+  gap: 0.65rem;
+  width: auto;
+  max-width: calc(100% - 2.5rem);
+  padding: 0.45rem 0.72rem 0.45rem 0.45rem;
+  border: 1px solid var(--pf-line);
+  border-radius: var(--pf-radius);
+  background: color-mix(in srgb, var(--pf-bg) 82%, transparent);
+  box-shadow: 0 18px 45px rgba(23, 23, 23, 0.12);
+  backdrop-filter: blur(14px);
+}
+
+.pf-ambassador-logo {
+  width: clamp(3.6rem, 7vw, 4.8rem);
+  aspect-ratio: 1;
+  padding: 0;
+  border: 0;
+  border-radius: 12px;
+  background: transparent;
+  cursor: pointer;
+  display: block;
+  transform-origin: center;
+  transition: filter 180ms ease;
+}
+
+.pf-ambassador-logo:hover,
+.pf-ambassador-logo:focus-visible {
+  filter: drop-shadow(0 16px 28px color-mix(in srgb, var(--pf-accent-3) 28%, transparent));
+}
+
+.pf-ambassador-logo:focus-visible {
+  outline: 2px solid var(--pf-accent-3);
+  outline-offset: 4px;
+}
+
+.pf-ambassador-logo.is-flying {
+  position: relative;
+  z-index: 20;
+  filter: drop-shadow(0 22px 34px color-mix(in srgb, var(--pf-accent-3) 34%, transparent));
+  will-change: transform;
+}
+
+.pf-ambassador-logo video,
+.pf-ambassador-logo img {
+  display: block;
+  width: 100%;
+  height: 100%;
+  border-radius: 12px;
+  object-fit: contain;
+  pointer-events: none;
+}
+
+.pf-ambassador-k {
+  display: block;
+  margin-bottom: 0.18rem;
+  color: var(--pf-accent-3);
+  font-size: 0.68rem;
+  font-weight: 850;
+  line-height: 1.1;
+  text-transform: uppercase;
+}
+
+.pf-ambassador-v {
+  display: block;
+  color: var(--pf-ink);
+  font-size: clamp(0.92rem, 1.5vw, 1.05rem);
+  font-weight: 800;
+  line-height: 1.18;
+}
+
+.pf-portrait-stage {
+  position: relative;
+  min-height: clamp(360px, 38vw, 450px);
+  border: 1px solid var(--pf-line);
+  border-radius: var(--pf-radius);
+  background:
+    linear-gradient(135deg, color-mix(in srgb, var(--pf-accent) 13%, transparent), transparent 46%),
+    linear-gradient(315deg, color-mix(in srgb, var(--pf-accent-2) 14%, transparent), transparent 46%),
+    var(--pf-surface);
+  overflow: hidden;
+  box-shadow: var(--pf-shadow);
+}
+
+.pf-portrait-stage::before {
+  content: "";
+  position: absolute;
+  inset: 0.85rem;
+  border: 1px solid var(--pf-line);
+  border-radius: 6px;
+  pointer-events: none;
+  z-index: 2;
+}
+
+.pf-portrait {
+  position: absolute;
+  inset: 0;
+  display: grid;
+  place-items: end center;
+  overflow: hidden;
+  background: var(--pf-surface-soft);
+}
+
+.pf-portrait img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  object-position: center bottom;
+  filter: saturate(0.92) contrast(1.03);
+  transform: none;
+  opacity: 1;
+  transition: opacity 260ms ease;
+}
+
+.pf-portrait img.is-switching {
+  opacity: 0.18;
+}
