+++ js/theme-boot.js
@@ -1,0 +1,12 @@
+(function () {
+  try {
+    var storageKey = 'mg-color-theme';
+    var storedTheme = localStorage.getItem(storageKey);
+    var prefersDark = window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches;
+    var theme = storedTheme || (prefersDark ? 'dark' : 'light');
+    document.documentElement.dataset.theme = theme;
+    document.documentElement.style.colorScheme = theme;
+  } catch (error) {
+    document.documentElement.dataset.theme = 'light';
+  }
+})();
