+++ project_details/project-budget-buddy.html
@@ -15,20 +15,7 @@
     <link
         href="https://fonts.googleapis.com/css2?family=Plus+Jakarta+Sans:wght@100;200;300;400;500;600;700;800;900&display=swap"
         rel="stylesheet" />
-    <script>
-        (function () {
-            try {
-                var storageKey = 'mg-color-theme';
-                var storedTheme = localStorage.getItem(storageKey);
-                var prefersDark = window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches;
-                var theme = storedTheme || (prefersDark ? 'dark' : 'light');
-                document.documentElement.dataset.theme = theme;
-                document.documentElement.style.colorScheme = theme;
-            } catch (error) {
-                document.documentElement.dataset.theme = 'light';
-            }
-        })();
-    </script>
+    <script src="../js/theme-boot.js"></script>
     <!-- Bootstrap icons-->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.1/font/bootstrap-icons.css" rel="stylesheet" />
     <!-- Core theme CSS (includes Bootstrap)-->
@@ -47,32 +34,7 @@
         </button>
     </div>
     <main class="flex-shrink-0">
-        <!-- Navigation-->
-        <nav class="navbar navbar-expand-lg py-3">
-            <div class="container px-5">
-                <a class="navbar-brand" href="../index.html">MG</a>
-                <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
-                    data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
-                    aria-expanded="false" aria-label="Toggle navigation">
-                    <span class="navbar-toggler-icon"></span>
-                </button>
-                <div class="collapse navbar-collapse" id="navbarSupportedContent">
-                    <div class="ms-auto d-flex align-items-center flex-column flex-lg-row gap-3">
-                        <ul class="navbar-nav mb-2 mb-lg-0 small fw-bolder text-center text-lg-start">
-                            <li class="nav-item"><a class="nav-link" href="../index.html">Home</a></li>
-                            <li class="nav-item"><a class="nav-link" href="../projects.html">Projects</a></li>
-                            <li class="nav-item"><a class="nav-link" href="../apps.html">Apps</a></li>
-                            <li class="nav-item"><a class="nav-link" href="../resume.html">Resume</a></li>
-                            <li class="nav-item"><a class="nav-link" href="../contact.html">Contact</a></li>
-                        </ul>
-                        <button class="theme-toggle-btn mt-2 mt-lg-0" type="button" data-theme-toggle
-                            aria-label="Toggle color scheme">
-                            <i class="bi bi-moon"></i>
-                        </button>
-                    </div>
-                </div>
-            </div>
-        </nav>
+        <nav class="navbar navbar-expand-lg" data-active-page="projects"></nav>
 
         <!-- Project Details Section-->
         <section class="py-5">
@@ -294,19 +256,10 @@
         </section>
     </main>
     <!-- Footer-->
-    <footer class="pf-footer py-4 mt-auto">
-        <div class="pf-shell d-flex flex-column flex-sm-row justify-content-between gap-2">
-            <div class="small">Copyright &copy; inventure71.github.io 2026</div>
-            <div class="small d-flex gap-3">
-                <a href="https://github.com/Inventure71" target="_blank" rel="noopener">GitHub</a>
-                <a href="https://www.linkedin.com/in/matteo-giorgetti-026172247/" target="_blank" rel="noopener">LinkedIn</a>
-                <a href="/assets/docs/matteo-giorgetti-resume-current.pdf">Resume PDF</a>
-            </div>
-        </div>
-    </footer>
+    <footer class="pf-footer py-4 mt-auto"></footer>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"></script>
     <script src="../js/theme.js"></script>
-    <script src="../js/playfolio.js"></script>
+    <script type="module" src="../js/playfolio.js"></script>
     <script src="../js/project-pages.js"></script>
 </body>
